Home menu changes

+++ HomeBuyer/HomeBuyer/Model/homeInfo.h
@@ -18,7 +18,7 @@ typedef enum
 @interface homeInfo : NSObject
 @property (nonatomic) homeType          mHomeType;
 @property (nonatomic, strong) NSString* mIdentifiyingHomeFeature;
-@property (nonatomic) uint              mHomeListPrice;
+@property (nonatomic) UInt64            mHomeListPrice;
 @property (nonatomic) uint              mHOAFees;
 @property (nonatomic) uint              mMonthlyLoanPayment;
 @property (nonatomic) homeAddress*      mHomeAddress;
